Committing some work done on a new way to build the simulation grid, which will end when restarting from files can be done. Note that this is work in progress, everyting should still work as previously.
parameters: Did a major revision here. Boost::program_options is
wrapped inside parameters, because I did not want to include it
everywhere. Classes can now add parameters they require to Parameters
via Parameters::add functions, ask Parameters to (re)parse the input
files, and then query the read values. In the future the classes should cache
the parameter values they need internally, instead of reading values
of static Parameter member variables (which are now deprecated).

gridbuilders now contains a couple new classes which can be used to build
the simulation grid in the future (when this work is complete). The
revamp of simulation grid building causes a lot of changes in pargrid
and elsewhere.