Two changes: parallel reading of parameters and introduction of wtime to measure walltime
The old class Parameters was split into two classes; Parameters which
contain constants, and Readparameters which reads in options from the
input file. Readoptions is able to read in options in parallel and thus
when using it one does not need to take care of broadcasting variables
etc.

In main and in timer.cpp the clock funciotns were replaced with
MPI__Wtime as it has better accuracy.