Two main updates:
Leveque and field solver boundary conditions are now set through user-specific
project files. Some simple boundary conditions are defined in the
corresponding projects/projects_*.h files. Acceleration at cell faces
is also calculated in the project files. While this is not necessary
for "normal" runs, it helps a lot to make automatic diagnostics/test
runs if the acceleration can be changed.

Second update is restart builder, i.e. a GridBuilder which reads a
.vlsv file and restarts the simulation using that file and the
corresponding .cfg file. This caused changes in many places, as the
GridBuilder interface needed some refining and all necessary data was
not written on the restart files.