Addition of the thermal pressure term as m/3*integral((V-V_avg)^2*f(r,v))d^3V. This closes JIRA issue #36.
Compiles and runs without crashes in all combinations of grid and solvers except
some cases of DCCRG+LEVEQUE, see JIRA issue #38.

The variable giving the number of cell paameters has been removed from common.h.
It has been replaced everywhere by CellParams::SIZE_CELLPARAMS, the last entry of
the namespace.

The pressure output after the first step is erroneous and there are marked differences
between the last but one and last steps. This is probably related to the calculation
of the spatial moments, I open a JIRA issue for that JIRA