Preliminary inclusion of topoloy optimization.
It is not enabled by default, enable -DCRAY_TOPOLOGY_OPTIMIZATION to
enable it for meteo.fmi.fi. To use it on other seastar Cray machines the
torus size also needs to be modified in the source code.

On meteo the speedup with leveque+harm1D is of the order of 20% for a
few hundred cores