Added some dpendencies for project
Now one can compile vlasiator in parallel. On Meteo the fastest way is
to compile on a disk visible to compute nodes ( wrk or stornext), and
use a command like this:

aprun -n 1 -d 12 make -j 12

This compiles it on an interactive node. One can of course also make a
batch script for this.