Found a bug that appeared sporadically for test_trans, even in serial mode.
When the unordered map in dccrg is filled with spatialcells, it not only
calls the constructor, but also the copy constructor. Apparently it
calls the constructor repeatedly on the same object, and copies it in
place. When using the default copy contructor the block_address_cache
vector contained garbage values, pointing all null_blocks to the same
null_block...

Now the serial test_trans works, hopefully reliably. Parallel still
broken




QUESPACE-87
QUESPACE-89
QUESPACE-90