QUESPACE-136
Added some functions, including average-subtracted calculations,
infinity-distance. The interface changed slightly as well, the functions take
two real pointer arguments in which they return the absolute and relative
value of the distance considered. Then this can be couted or written to file
or whatever one wishes to do with them.