Fixes QUESPACE-137
Now we do not flush all data when we take into use a new blocklist that
has been received over MPI. Instead we just add and remove the blocks
that need to be added or removed.

USe a separate blocklist for MPI communication, now the cell state is
always well defined, previosuly it sould be in a state where the
blocklist did not correspond to the actual blocks in the cell.

Fixes some compiler warnings in spatialcell class