e-organized how parameters are read in.
All parameters are now visible using the --help option

Projects need to define addProjectParameter and getProjectParameter.
These were now implemented in Fluctutations, but not yet in the others.

Both the readparameters class and the parameters struct were in the same
file. Now these are in their own set of files.

Moved the add and get operations of the global parameters in the
Parameters struct to struct specific functions. The readConfigFile
functions is removed from vlasiator.cpp. (towards QUESPACE-104 )