Fixes QUESPACE-204 and QUESPACE-201
Now file writes are defined by intervals in simualted seconds, not
steps. The new variables are 
  --system_write_t_interval arg (=1.797693134862316e+308)
                                        Save the simulation every arg simulated seconds
  --restart_write_t_interval arg (=1.797693134862316e+308)
                                        Save the complete simulation every arg simulated seconds

By default none of the files are written out ever.
 

If dynamic timestepping is on, then the timesteps are adjusted so
that the exact interval is hit. If dt is fixed, then files are written
out once the time hits, or is past the requested interval. An epsilon of
1ps is included in the tests to avoid most floating point comparison
problems.

files with reduced data are caled grid.xxx.vlsv, where xxx is a running
index of files written out (not simulation step)

files with full data are called restar.xxx.vlsv, where xxx is the
corresponding running index for restart files.


Also removed unused variables, and an unused function
(calculateSimParameters)

dynamic timestepping is now default