QUESPACE-206
Simplified the interface for the boundary conditions for derivatives so that
the user can set the derivatives by hand in the
project.h:fieldSolverBoundaryCondDeriv[XYZ] functions directly.

In test cases with periodic boundaries the functions just return. In cases
with non-periodic boundaries the code has been added to set the
derivatives. Only in Harris.h are the values actually different from 0.0.