QUESPACE-220
This update effectively removes all references to the FACE components of
fields. All that remains are the volume averages and the B face and E edge
components.

The field solver is now designed such that it takes care of computing the
volume averages *even when the field solver is turned off*, i.e. the user only
has to set the (face) B components initially in any case that may occur.

In the wake of the latter changes the field solver functions do not check
anymore whether the fields are to be propagated, as they only get called by
vlasiator.cpp:main() if it is the case (and to initialise the fields from the
face B components when the field solver is turned off, obviously).

Note that with respect to QUESPACE-50 the BnFACEn0 components have been
renamed to Bn0 throughout the code as well including in the projects Fluctuations_B0 and test_fp_B0, caveat emptor.