QUESPACE-169
Commit wrapping up what was left open in previous commits of this
branch.

The folder structure of sysboundary and datareduction has also been submitted,
it wasn't there in previous commits...

The sysBoundaryFlag is now an uint in the SpatialCell class.

The code has been made compliant to the convention of having system boundaries
(prefix sys in names) and process boundaries (prefix proc) to avoid confusion
between boundaries, ghost cells etc.

At this stage the code runs (at least initialises, more hasn't been tested as
it is useless at the moment) and assigns the sysBoundaryFlag correctly. In
some places computation is bypassed for cells flagged as sysboundarytype::DO_NOT_COMPUTE.