QUESPACE-169
Cells are being initialised according to their sysBoundaryFlag. However the
functions for the cases where the fields or the distribution function have to
differ from the general case (i.e. solar wind) have to be filled with sensible
code (importing from file instead of zeroing).

The precedence of the various system boundary types is imported from the cfg
file and handled through a map.

Weights are set to 0 for DO_NOT_COMPUTE cells.