QUESPACE-229
Now making use of the bool return of the Readparameters::get() functions to
check for consistency between Readparameters::add() and Readparameters::get()
in some places (system boundary conditions and Magnetosphere project). It
makes the code substantially heavier to get a large number of parameters thus
it is not done in parameters.cpp.