QUESPACE-169 QUESPACE-226
Patched the system boundary conditions to set the intermediate moments of
leapfrogging. This effectively only supports constant conditions as e.g. RHO
and RHO_DT2 are set to the same value. In addition once more derivatives are
set to 0 for want of a better solution for now.

QUESPACE-281

This version implements the E_ION (eq. 3) plugged into (eq. 2) in Malakit et
al. GRL 36, 2009 and replaces the use of E_VOL in the Lorentz force. It has
been tested on a dispersion case which looks marginally better.