QUESPACE-232
Beginning of transfer of projects to the class structure. Preliminary tests
with the Magnetosphere test are showing no significant change of performance.

Projects ported to the new interface so far: Alfven, Diffusion, Dispersion,
Magnetosphere.

Note: the cfg has to have an entry project = XXX where XXX is what you want to
run or the default will kick in (Fluctuations at the moment).

Dispersion has been modified to make the RNG things class members and it has
been tested such as to be thread-safe by putting the random values needed in
sevral places as threadprivate (OpenMP) static private (C++) class members.

Makefile changes are of course also done, the executable does not bear the
name of the project as all are compiled into it. One could question the
usefulness of keeping the architecture and floating-point precision
indications in the name.