QUESPACE-232 QUESPACE-299
First submission and testing of a scheme to initialise isotropic maxwellians
only in their vicinity and not by sampling the whole (potentially vast and
mostly void) velocity space.

Has also been done for the SetMaxwellian inflow system boundary condition and
that gives identical results so it seems sane. Note that for the inflow this
is now default and cannot be changed easily whereas for projects this is only
added to Magnetosphere so far. Basically one has to inherit the project from
the IsotropicMaxwellian class and move the uniform bulk velocity data class
member to the IsotropicMaxwellian class and (re)name it Real V0[3].