Fixed incorrect spatial cell id search.
Related to issue #81

vlsvextract was reading cell IDs from the array MESH, which contains
the global IDs for both local and ghost blocks, but did not take into
account the existence of ghost cells. This caused variables to be read
from wrong cells. Reverted back to the old behaviour, i.e., to read
the cell IDs from a VARIABLE called CellID.