Changed Makefile -DPAPI_MEM definition.
It is now defined in Makefile only if it has not been already defined
in architecture-specific Makefile. If you set

PAPI_FLAG=

in your architecture-specific Makefile, then "-DPAPI_MEM" is not added
to CXXFLAGS in Makefile. This makes it more convenient to develop the
code in environments without papi, e.g, home laptop.

Conflicts:
	MAKE/Makefile.arto