Added phiprof profiling to write functions
This commit adds more phiprof profiling to writeGrid and writeRestart. It also added the GB written as a work unit.

To make more sensible profiling information it also added a barrier in the beginning of the functions. Without this the load imbalance shows up in the vlsvwrite open call, as that makes collective operations.