Particle Pusher: allow configurable input frequency.
Before, it was hardcodedly assumed that bulk input files were given as
one per second, which is entirely untrue for newer runs.
This can now be configured with the particles.input_dt parameter.

(Actually only affects the first file to be opened, thereafter the
actual time / t parameters of subsequent files are checked)