Multipop-enabled IPShock project.
Shock width can only be specified globally, and not per-population,
since it is required for the field transition calculation.
Im not fully sure if this is the proper way to go here (since different
species may very well have different shock transition lengths, thus
leading to a cross shock potential), but it was the easiest fix that
kept the code basically intact.

Also, this is now basically untested... it compiles though!