Fix particle alignment bug by handing a custom allocator to std::vector.
The C++ Standard Library containers ignore alignment specifiers on datatypes,
and just slap them in memory whereever they want.

At the same time, GCC sees the alignment specifier of the Agner vectorclass
datatypes and stiffly assumes them to be aligned, thus using the VMOVAPD
instruction to load them from aligned memory addresses into the AVX registers.
This causes a bus error if the alignment turns out not to be correct.

The workaround here is a bit nasty, since it involves a custom allocator, but
seems to work ok.