Fix outflow and ionosphere boundaries to get symmetric B evaluation.
Before, B vectors were always copied in the sysboundary cells, which
caused the Yee-Lattice locations of these components to be asymmetric
wrt. the boundary. Now, it is checking whether the neighbour on the
corresponding components' side is a non-sysboundary and solving the
field normally if it is.

This should increase inner boundary accuracy. Hopefully. Testing in
progress.