Cut-cell precipitation DRO
As yet another alternative to the line-cutting precipitation DRO in
PR#670, this version implements an implementation where each velocity
cell is coverage-tested against the loss cone, allowing for partial
coverage to contribute partially.

The concrete implementation defines the loss cone as a signed distance
function, which returns the distance to the cone surface for every point
in vspace. If this distance is much larger than the vspace cell's
diagonal, the cell is assumed to be clearly outside. If it is negative
by the same distance, it is clearly inside.
But cells that are closer to the cone boundary then their diagonal are
recursively octree-subdivided, and each subcell is again tested in the
same way (up to a hardcoded recursion limit of 6).