Automatic compilation and testpackage run (for Continuous Integration).
Github actions automatically get triggered on every PR and commit for
the master and dev branches. Basic compilation tests are run on github
hosted runners (using a docker VM), testpackage is built and run on the
Carrington cluster.

Any newly produced warnings or errors as part of the PR code are
highlighted in the PR diff view. Testpackage results are available as a
PR check (see "details" for the whole output).