CI: Don't upload testpackage_build.log anymore, get proper return val.
Before, the pipe catching the compilation output was causing the job to
always return a green checkmark, even if the build failed for some
reason.

The information in that build log is anyway captured as part of the
github action, so there's no need to keep it around as an artefact.