add "BOUNDARY_PADDING" boundarytype, for non-coupled DO_NOT_COMPUTE.
This type only gets assigned to fsgrid cells that are left outside the
actual simulation domain after dccrg refinement. They should not be
computed or coupled into dccrg, as not to distort the volume averages.

Not that there is special treatment near the ionosphere, where *do* want
to couple fsgrid cells.