Pre-commit config file for whitespace and formatting fixes
This commit contains a config file for pre-commit, which automatically
adjusts whitespace and formatting for source files. It *also* contains
all the changes this config would enact onto the source tree, fixing all
trailing whitespace problems and tab/space conversion in the source.
This is a big mess, so I suggest delaying this PR and re-creating it at
some big merge point.

Note that the pre-commit config also contains a commented-out hook for
clang-format according to the clang-format config file already in the
repository, but I consider it too aggressive to automatically enforce it
everywhere.